remove/combine unused util classes
