+++ src/main/java/com/lothrazar/cyclicmagic/ModMain.java
@@ -161,6 +161,8 @@ public class ModMain {
 	}
 	/* 
 	 * TODO LIST
+	 * 
+	 * SORTING
 	 *  
 	 * heart container-PERSIST ON DEATH
 	 *
