+++ src/main/java/com/lothrazar/cyclicmagic/gui/spell/ButtonSpellToggle.java
@@ -5,7 +5,7 @@ import java.util.List;
 
 import com.lothrazar.cyclicmagic.gui.button.ITooltipButton;
 import com.lothrazar.cyclicmagic.spell.ISpell;
-import com.lothrazar.cyclicmagic.util.UtilString;
+import com.lothrazar.cyclicmagic.util.UtilChat;
 import com.lothrazar.cyclicmagic.util.UtilTextureRender;
 
 import net.minecraft.client.Minecraft;
@@ -68,7 +68,7 @@ public class ButtonSpellToggle extends GuiButton implements ITooltipButton {
 
 		// tooltips.add(EnumChatFormatting.LIGHT_PURPLE + spell.getInfo());
 
-		tooltips.addAll(UtilString.splitIntoLine(spell.getInfo(), 28));
+		tooltips.addAll(UtilChat.splitIntoLine(spell.getInfo(), 28));
 
 		// boolean unlocked =
 		// ItemCyclicWand.Spells.isSpellUnlocked(props.getPlayer().getHeldItem(),
