+++ src/main/java/com/lothrazar/cyclicmagic/item/ItemFoodHeart.java
@@ -1,7 +1,7 @@
 package com.lothrazar.cyclicmagic.item;
 
-import com.lothrazar.cyclicmagic.IHasConfig;
 import com.lothrazar.cyclicmagic.IHasRecipe;
+import com.lothrazar.cyclicmagic.registry.CapabilityRegistry;
 import com.lothrazar.cyclicmagic.util.Const;
 import com.lothrazar.cyclicmagic.util.UtilEntity;
 
@@ -10,10 +10,9 @@ import net.minecraft.init.Items;
 import net.minecraft.item.ItemFood;
 import net.minecraft.item.ItemStack;
 import net.minecraft.world.World;
-import net.minecraftforge.common.config.Configuration;
 import net.minecraftforge.fml.common.registry.GameRegistry;
 
-public class ItemFoodHeart extends ItemFood implements IHasRecipe,IHasConfig{
+public class ItemFoodHeart extends ItemFood implements IHasRecipe{
  
 	private static final int numFood = 2;
 	private static final int numHearts = 1;
@@ -26,14 +25,10 @@ public class ItemFoodHeart extends ItemFood implements IHasRecipe,IHasConfig{
 	@Override
 	protected void onFoodEaten(ItemStack par1ItemStack, World world, EntityPlayer player) {
  
+//		IPlayerExtendedProperties prop = CapabilityRegistry.getPlayerProperties(player);
+		
 		//one heart is 2 health points (half heart = 1 health)
 		UtilEntity.incrementMaxHealth(player, 2 * numHearts);  
-	}
-//Edible hearts that increase your heath (permanently; until death)
-	@Override
-	public void syncConfig(Configuration config) {
-
-//TODO: get hearts persisting through death config to wipe on death?
 	}
 
 	@Override
