+++ src/main/java/com/lothrazar/cyclicmagic/util/UtilInventory.java
@@ -186,5 +186,9 @@ public class UtilInventory {
 			entityPlayer.inventory.decrStackSize(currentItem, 1);
 		}
 	}
+	
+	
+	
+	
 
 }
