+++ src/main/java/com/lothrazar/cyclicmagic/util/UtilSound.java
@@ -2,7 +2,6 @@ package com.lothrazar.cyclicmagic.util;
 
 import net.minecraft.block.Block;
 import net.minecraft.entity.EntityLivingBase;
-import net.minecraft.entity.passive.EntityVillager;
 import net.minecraft.entity.player.EntityPlayer;
 import net.minecraft.util.SoundCategory;
 import net.minecraft.util.SoundEvent;
